% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_functions.R
\name{gd_get_biome}
\alias{gd_get_biome}
\title{Get the biome, temperature and precipitation of a site}
\usage{
gd_get_biome(data, si_lat, si_long, si_mat, si_map, merge_deserts = FALSE)
}
\arguments{
\item{data}{Data frame including either geographical coordinates (latitude
and longitude) or climatic data (MAT and MAP).}

\item{si_lat}{Character, name of the column containing latitude values in
decimal degrees or, alternatively, if \code{data} not provided, a numeric
value of latitude.}

\item{si_long}{Character, name of the column containing longitude values in
decimal degrees or, alternatively, if \code{data} not provided, a numeric
value of longitude.}

\item{si_mat}{Character, name of the column containing mean annual
temperature values in degrees Celsius or, alternatively, if \code{data}
not provided, a numeric value of MAT.}

\item{si_map}{Character, name of the column containing mean annual
precipitation values in millimeters or alternatively, if \code{data} not
provided, a numeric value of MAP.}

\item{merge_deserts}{Logical indicating if desert biomes should be merged
in a single biome. By default, deserts are not merged.}
}
\value{
A data frame with the original data and extra columns of
   mean annual temperature, mean annual precipitation and biome.
}
\description{
\code{gd_get_biome} takes either geographical coordinats or climatic data
and returns the Whittaker's biome.
}
\details{
This function takes a data frame including at least either two columns
with geographical coordinates (latitude and longitude), or two columns with
climatic data (mean annual temperature and mean annual precipitation).
It returns the same data frame with extra columns of climatic data (si_mat
and si_map, if not provided) and biome (si_biome), according to
\code{\link{gd_get_biomes_spdf}}.

Alternatively it accepts either a value of latitude and a value of longitude
or a value of mean annual temperature (MAT) and a value of mean annual
precipitation (MAP), and returns a data frame with latitude (if provided),
longitude (if provided), mean annual temperature, mean annual precipitation
and biome.

Climatic data is obtained from WorldClim 1.4 using the \code{RFc} package.
}
\examples{
# Create a data frame with coordinates
sites <- data.frame(site = c("A", "B"), lat = c(69.49, 41.43),
                    long = c(27.23, 2.07))

# Get temperature, precipitation and biomes
gd_get_biome(sites, si_lat = "lat", si_long = "long")

# Alternatively, for one site
gd_get_biome(si_mat = 15.8, si_map = 622)

}
\seealso{
Other Get.Data.Functions: \code{\link{gd_get_biomes_spdf}}
}

