% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_functions.R
\name{gd_get_biomes_spdf}
\alias{gd_get_biomes_spdf}
\title{Make a SpatialPolygonsDataFrame object of the biomes}
\usage{
gd_get_biomes_spdf(merge_deserts = FALSE)
}
\arguments{
\item{merge_deserts}{Logical indicating if desert biomes should be merged
in a single biome. By default, deserts are not merged.}
}
\value{
An object of class SpatialPolygonsDataFrame
}
\description{
\code{gd_get_biomes_spdf} returns a SpatialPolygonsDataFrame object of
the Whittaker's biomes modified by Ricklefs (2008) in function of mean
annual temperature (MAT) and mean annual precipitation (MAP)
(MAT in degrees Celsius and MAP in mm).
}
\details{
This function is not commonly used alone, but it is called by
\code{\link{gd_get_biome}}.
}
\examples{
# Get the SpatialPolygonsDataFrame object of the biomes
biomes_spdf <- gd_get_biomes_spdf()

}
\seealso{
Other Get.Data.Functions: \code{\link{gd_get_biome}}
}

