% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_functions.R
\name{vis_biome}
\alias{vis_biome}
\title{Plot a diagram of biomes}
\usage{
vis_biome(merge_deserts = FALSE)
}
\arguments{
\item{merge_deserts}{Logical indicating if desert biomes should be merged
in a single biome. By default, deserts are not merged.}
}
\value{
A ggplot object showing the biomes.
}
\description{
\code{vis_biome} produces a ggplot object showing the Whittaker's biomes as
colored areas according to mean annual temperature (MAT) and mean annual
precipitation (MAP) using a SpatialPolygonsDataFrame object obtained with
\code{\link{gd_get_biomes_spdf}}
}
\examples{
# Create a ggplot diagram of the biomes
vis_biome()

}
\seealso{
Other Visualization.Functions: \code{\link{vis_location_biome}}
}

