% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_functions.R
\name{vis_location_biome}
\alias{vis_location_biome}
\title{Plot a diagram of biomes with sites as dots}
\usage{
vis_location_biome(data, si_lat, si_long, si_mat, si_map, point_labels,
  col_fill = "white", merge_deserts = FALSE)
}
\arguments{
\item{data}{Data frame including either geographical coordinates (latitude
and longitude) or climatic data (MAT and MAP).}

\item{si_lat}{Character, name of the column containing latitude values in
decimal degrees or, alternatively, if \code{data} not provided, a numeric
value of latitude.}

\item{si_long}{Character, name of the column containing longitude values in
decimal degrees or, alternatively, if \code{data} not provided, a numeric
value of longitude.}

\item{si_mat}{Character, name of the column containing mean annual
temperature values in degrees Celsius or, alternatively, if \code{data}
not provided, a numeric value of MAT.}

\item{si_map}{Character, name of the column containing mean annual
precipitation values in millimeters or alternatively, if \code{data} not
provided, a numeric value of MAP.}

\item{point_labels}{A character vector the same length of the number of
sites to be plotted, indicating site labels. If missing, no labels are
plotted.}

\item{col_fill}{A character vector the same length of the number of
sites to be plotted, indicating the fill color for site points, or
alternatively a character vector of length one indicating a single
fill color for all sites. The default is white.}

\item{merge_deserts}{Logical indicating if desert biomes should be merged
in a single biome. By default, deserts are not merged.}
}
\value{
A ggplot object showing the biomes and sites.
}
\description{
\code{vis_location_biome} produces a ggplot object showing the biomes as
colored areas according to mean annual temperature (MAT) and mean annual
precipitation (MAP), using the function \code{\link{vis_biome}}, and adds
the sites on it according to their values of MAT and MAP, obtained using
the function \code{\link{gd_get_biome}} if not provided.
}
\examples{
# Create a data frame with coordinates
sites <- data.frame(site = c("A", "B"), lat = c(69.49, 41.43),
                    long = c(27.23, 2.07))

# Create a ggplot diagram of the biomes with sites
vis_location_biome(sites, si_lat = "lat", si_long = "long",
                   point_labels = sites$site)

# Alternatively, for one site
vis_location_biome(si_mat = 15.8, si_map = 622, col_fill = 'red')

}
\seealso{
Other Visualization.Functions: \code{\link{vis_biome}}
}

